[android] Moved some util methods.
