+++ android/src/com/mapswithme/util/UiUtils.java
@@ -5,7 +5,6 @@ import android.app.Activity;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.res.Configuration;
-import android.net.Uri;
 import android.os.Build;
 import android.provider.Settings;
 import android.support.annotation.DimenRes;
@@ -22,6 +21,7 @@ import android.view.animation.Animation;
 import android.view.animation.Animation.AnimationListener;
 import android.view.animation.AnimationUtils;
 import android.widget.TextView;
+
 import com.mapswithme.maps.MwmApplication;
 import com.mapswithme.maps.R;
 
@@ -169,38 +169,6 @@ public final class UiUtils
   {
     tv.setText(text);
     showIf(!TextUtils.isEmpty(text), tv);
-  }
-
-  public static void openAppInMarket(Activity activity, String marketUrl)
-  {
-    try
-    {
-      activity.startActivity(new Intent(Intent.ACTION_VIEW, Uri.parse(marketUrl)));
-    } catch (final Exception e)
-    {
-      e.printStackTrace();
-    }
-  }
-
-  public static void showFacebookPage(Activity activity)
-  {
-    try
-    {
-      // Exception is thrown if we don't have installed Facebook application.
-      activity.getPackageManager().getPackageInfo(Constants.Package.FB_PACKAGE, 0);
-
-      activity.startActivity(new Intent(Intent.ACTION_VIEW, Uri.parse(Constants.Url.FB_MAPSME_COMMUNITY_NATIVE)));
-    } catch (final Exception e)
-    {
-      activity.startActivity(new Intent(Intent.ACTION_VIEW, Uri.parse(Constants.Url.FB_MAPSME_COMMUNITY_HTTP)));
-    }
-  }
-
-  public static void showTwitterPage(Activity activity)
-  {
-    Intent intent;
-    intent = new Intent(Intent.ACTION_VIEW, Uri.parse(Constants.Url.TWITTER_MAPSME_HTTP));
-    activity.startActivity(intent);
   }
 
   public static void checkConnectionAndShowAlert(final Activity activity, final String message)
