+++ android/src/com/mapswithme/util/Utils.java
@@ -264,6 +264,24 @@ public class Utils
     }
   }
 
+  public static void showFacebookPage(Activity activity)
+  {
+    try
+    {
+      // Exception is thrown if we don't have installed Facebook application.
+      activity.getPackageManager().getPackageInfo(Constants.Package.FB_PACKAGE, 0);
+      activity.startActivity(new Intent(Intent.ACTION_VIEW, Uri.parse(Constants.Url.FB_MAPSME_COMMUNITY_NATIVE)));
+    } catch (final Exception e)
+    {
+      activity.startActivity(new Intent(Intent.ACTION_VIEW, Uri.parse(Constants.Url.FB_MAPSME_COMMUNITY_HTTP)));
+    }
+  }
+
+  public static void showTwitterPage(Activity activity)
+  {
+    activity.startActivity(new Intent(Intent.ACTION_VIEW, Uri.parse(Constants.Url.TWITTER_MAPSME_HTTP)));
+  }
+
   public static void sendSupportMail(Activity activity, String subject)
   {
     final Intent intent = new Intent(Intent.ACTION_SEND);
